requirements.txt
-e svn+http://code.djangoproject.com/svn/django/trunk#egg=Django

django-extensions
django-debug-toolbar # -e git+git://github.com/robhudson/django-debug-toolbar.git#django-debug-toolbar
django-reversion
django-annoying
django-beehat
docutils
feedparser
ipython
lxml
monit
psycopg2
south
Werkzeug
